backend/requirements.txt
anyio==3.6.2
certifi==2022.12.7
chardet==3.0.4
click==8.1.3
fastapi==0.95.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httptools==0.5.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
